Modified classes to improve understandability
Changed the AtomImageConstants class so that the default pixel size
is now 6.45 um, which is apparently the value for the gravimenter.
Also changed the width/gamma value to be in MHz rather than s^{-1}

Changed the AbsorptionImage class to have both image and an
imageCorr properties.  The image property is the raw OD, and the
imageCorr property is the corrected OD including saturation,
polarization, and detuning.  "image" is used for display, while
"imageCorr" is used for calculations.  Method by which number is
calculated has been updated to match the calculation of imageCorr

Abs_Analysis has been updated to reflect the above changes.