Updates to fit functions and parameter estimation
Updated AtomCloudFit BEC profile so that it is in line with the 2D
cloud, so that the fit parameter for the amplitude of the profile
is the same as the actual OD.  Also added linear slopes to the
profile

Updated AtomCloudFit to have a 'none' fit type.

Updated AbsorptionImage to calculate absorption image as only the
real part of the optical depth.

Updated CloudParameters to work properly with no input arguments