Added background removal, sum calc to AbsorptionImage
AbsorptionImage.m now has a background removal function which
uses a least-squares minimization routine to match the background
image to the with-atoms image.  It excludes the regions where there
are atoms.

Also conatins a "fitting" method for just summing up OD data to
get a value for the number of atoms, rather than fitting for it

Finally, changes to the plot functions so that they are compatible
with multiple ROIs