Can now reliably save AbsorptionImage class and related classes
Added struct(), saveobj(), and loadobj() methods to all important
imaging analysis classes (neglecting the error class) so that
data can be saved without taking up huge amounts of space.

Processed data, such as the raw and corrected OD maps, are not
saved. Instead, the raw image data is saved in uint16 format, and
the OD maps are recreated on load.