Fixed bug with copying RawImageData objects
Abs_Analysis_NClouds was not copying the raw image data properly.
This was because the way that the cloud variable was created meant
all the raw data objects were in fact the same object.  Changed how
this was done to avoid the problem

Also changed the copying behaviour of RawImageData to properly copy
the status property, which is an instance of handle class
ImageAnalysisErrorHandler.