Major back-end changes to AtomCloudFit and CloudParameters
Changed how fits are handled in AtomCloudFit.  First, the user can
now supply upper and lower bounds and a guess for any parameter via
the new properties ub, lb, and guess, which are of type
CloudParameters.  Second, 1D and 2D fits are now handled by two
functions fit1D and fit2D.  Different fit types are handled by
specifying an "order" cell array which describes the order in which
the fit parameters occur in the array used by lsqcurvefit.  The
class CloudParameters has been changed to account for this, with new
methods compare(), convert2array(), and convertFromArray().

CloudParameters has now been changed to a handle class for convenience.

Finally, 1D and 2D backgrounds have been systematized for 1D and 2D
fits. It is assumes that the last 2 (3) parameters from the fit
correspond to the offset and linear terms for 1D (2D) fits.
Additionally, the background data is saved as a new property in
AtomCloudFit so that we can use it to accurately compute offsets.
This eliminates the need to know how the background was referenced.