Modified reporting of peak OD, handling of NaNs
Peak OD is now reported based on the corrected OD divided by the
polarization correction so that it is closer to what the peak raw
OD actually is.

Previously, if the OD was high enough such that all the light was
absorbed and intensity - dark image was less than zero, the OD was
set to 0.  Instead, the OD is now set to NaN.  Fitting seems to still
work.

Fixed bug in BinaryImageData when loading mutliple image sets.