Fixed bugs in get_trap_freq and BinaryImageData
BinaryImageData parsed data wrong because it started parsing at the
wrong index.  This made Mono8 images look like there was a weird
line on the left side of the image.  Bug is now fixed.

get_trap_freq needed to be updated to compensate for changes to
optical_trap.m in matlab-common-functions